backend/requirements.txt
asgiref==3.11.1
attrs==25.4.0
diff-match-patch==20241021
Django==6.0.3
django-axes==8.3.1
django-cors-headers==4.9.0
django-debug-toolbar==6.2.0
django-extensions==4.1
django-import-export==4.4.0
django-simple-history==3.11.0
django-unfold==0.85.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
freezegun==1.5.5
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
model-bakery==1.23.3
packaging==26.0
pluggy==1.6.0
psycopg==3.3.3
psycopg-binary==3.3.3
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
tablib==3.9.0
typing_extensions==4.15.0
uritemplate==4.2.0
